Makefile|install.sh: Break up install.sh into scripts generated with make vars
The ./install.sh has been problematic as a lot of the configuration variables
defined within (and overridden with .local-conf) have also been required when
configuring the client & server (e.g. generating index.html).

Instead, make the variables part of the Makefile, and generate deploy scripts
for each part in the same manner as generating index.html

This means that any reconfiguring should also mean that "make" is re-run, but
this has been the case for a while, just not obviously.