server/app: Don't put connections back early when streaming output
Once Response() object is returned, the cur gets returned to the pool, but the
concordance search may not be done with it yet. Instead use flask.g to store a
per-request cursor, and stream_with_context() to make sure this is only closed
once we're done.

Given this, we can now close the cursor properly on put_cursor()