concordance: Move unidecode normalisation of results clientside
The KWICGrouper needs to go back from tokens -> lower-case types, so it can do
it's own searching. Before this operation was bodged by the server
unidecode-ing the returned tokens, then the client lower-casing them.

Instead the server should return unadulterated tokens, and the client should be
turning the tokens back into types for KWICGroup matching.

Ideally the API would be changed so it returns the type for each token, but that
would break backwards compatibility and require a fair few changes.