server/tokenizer: Add custom boundary-types so tokens include hyphens and apostrophes
It is more useful for us to have over-the-top as one word, and ignore the false
positives.

Apostrophes at extremities not being included when they are mid-word, e.g.
can’t, feels like a bug. In addition, if we handle them here it helps us ignore
them when parsing quotes.