server/region/suspension: Redefine suspensions
Paying attention to sentence boundaries at the end of the previous quote was
causing too many false positives, many quotes end with an exclamation, but no
following upper-case character, e.g.

    “My dear Miss Flora!” began Miss Rich

in CLiC 1.6 the exclamation mark wouldn't have been marked as a sentence
boundary (and thus would be considered a suspension). However the unicode
sentence boundaries consider this to be a boundary.

Instead of looking at the sentence boundaries, look at the first letter to make
sure it isn't upper-case. This will cuase valid suspensions to be dropped, e.g.

    ‘No, please go on!’ Alice said very humbly; ‘I won’t interrupt again.

...but there is no systematic ways to detect this vs. a genuine
start-of-sentence.