Move TaggedTextRegionMarkup/TaggedTextTableMarkup into taggedtext.py
It makes more sense for these classes to be documented on the same page
as the rest of TaggedText, so to do this move their sources into the
main module dir.

Keep markup/table.py for the actual printing algorithms so
taggedtext.py doesn't get too out of control, but these are now
entirely private.